+++ util/Validator/index.js
@@ -2,13 +2,15 @@
  * 校验器
  *
  * @method collect 用来收集基础表单组件的校验方法，通常为基础表单组件的 collectValidate 属性的值。
- *         注意，使用使一定要 bind 校验器示例。
  * @method run 用来运行结果，返回一个对象。
  *         如果运行正确，返回对象的 `err` 属性值为 0，返回对象的 `data` 属性值为包含所有收集表单的名字和值的映射；
  *         如果运行错误，返回对象的 `err` 属性值为 1，返回对象的 `data` 属性值为出错信息。
  */
 
 class Validator {
+  constructor() {
+    this.collect = this.collect.bind(this);
+  }
   validateList = [];
   collect(validate) {
     this.validateList.push(validate);